+++ src/Where.h
@@ -9,7 +9,6 @@
 #include <concepts>
 #include <unordered_map>
 #include <fmt/format.h>
-#include "MemberPointerUtils.h"
 #include "Reflect.h"
 #include "StringUtils.h"
